# backend/app/controllers/auth_controller.go
package controllers

import (
	"context"
	"encoding/json"
	"fmt"
	"os"
	"strconv"
	"time"

	"gitlab.xipat.com/omega-team3/qr-menu-backend/app/models"
	"gitlab.xipat.com/omega-team3/qr-menu-backend/pkg/repository"
	"gitlab.xipat.com/omega-team3/qr-menu-backend/pkg/utils"
	"gitlab.xipat.com/omega-team3/qr-menu-backend/platform/cache"
	"gitlab.xipat.com/omega-team3/qr-menu-backend/platform/database"
	"gorm.io/gorm/clause"

	"github.com/gofiber/fiber/v2"
	"gopkg.in/gomail.v2"
)

// UserSignUp method to create a new user.
// @Description Create a new user.
// @Summary create a new user
// @Tags User
// @Accept json
// @Produce json
// @Param email body string true "Email"
// @Param password body string true "Password"
// @Param user_role body string true "User role"
// @Success 200 {object} models.User
// @Router /v1/user/sign/up [post]
func UserSignUp(c *fiber.Ctx) error {
	// Create a new user auth struct.
	signUp := &models.SignUp{}

	// Checking received data from JSON body.
	if err := c.BodyParser(signUp); err != nil {
		// Return status 400 and error message.
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	// Create a new validator for a User model.
	validate := utils.NewValidator()

	// Validate sign up fields.
	if err := validate.Struct(signUp); err != nil {
		// Return, if some fields are not valid.
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   utils.ValidatorErrors(err),
		})
	}

	// Checking role from sign up data.
	role, err := utils.VerifyRole(signUp.UserRole)
	if err != nil {
		// Return status 400 and error message.
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	// Create a new user struct.
	user := &models.User{}

	// Set initialized default data for user:
	user.CreatedAt = time.Now()
	user.Email = signUp.Email
	user.PasswordHash = utils.GeneratePassword(signUp.Password)
	user.UserStatus = 1 // 0 == blocked, 1 == active
	user.UserRole = role

	// Validate user fields.
	if err := validate.Struct(user); err != nil {
		// Return, if some fields are not valid.
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   utils.ValidatorErrors(err),
		})
	}

	// Create a new user with validated data.
	if tx := database.Database.Create(&user); tx.Error != nil {
		// Return status 500 and create user process error.
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   "account is existed",
		})
	}

	// Delete password hash field from JSON view.
	user.PasswordHash = ""

	// Return status 200 OK.
	return c.JSON(fiber.Map{
		"error": false,
		"msg":   nil,
		"user":  user,
	})
}

// UserSignIn method to auth user and return access and refresh tokens.
// @Description Auth user and return access and refresh token.
// @Summary auth user and return access and refresh token
// @Tags User
// @Accept json
// @Produce json
// @Param email body string true "User Email"
// @Param password body string true "User Password"
// @Success 200 {string} status "ok"
// @Router /v1/user/sign/in [post]
func UserSignIn(c *fiber.Ctx) error {
	// Create a new user auth struct.
	signIn := &models.SignIn{}
	var foundedUser models.User

	// Checking received data from JSON body.
	if err := c.BodyParser(signIn); err != nil {
		// Return status 400 and error message.
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	// Get user by email.
	tx := database.Database.First(&foundedUser, "email = ?", signIn.Email)
	if tx.Error != nil {
		// Return, if user not found.
		return c.Status(fiber.StatusNotFound).JSON(fiber.Map{
			"error": true,
			"msg":   tx.Error.Error(),
		})
	}

	// Compare given user password with stored in found user.
	compareUserPassword := utils.ComparePasswords(foundedUser.PasswordHash, signIn.Password)
	if !compareUserPassword {
		// Return, if password is not compare to stored in database.
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   "wrong user email address or password",
		})
	}

	// Get role credentials from founded user.
	credentials, err := utils.GetCredentialsByRole(foundedUser.UserRole)
	if err != nil {
		// Return status 400 and error message.
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	strFoundedUserId := strconv.FormatUint(foundedUser.ID, 10)

	// Generate a new pair of access and refresh tokens.
	tokens, err := utils.GenerateNewTokens(strFoundedUserId, credentials)
	if err != nil {
		// Return status 500 and token generation error.
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	// Save refresh token to Redis.
	errSaveToRedis := cache.RedisClient.Set(context.Background(), strFoundedUserId, tokens.Refresh, 0).Err()
	if errSaveToRedis != nil {
		// Return status 500 and Redis connection error.
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   errSaveToRedis.Error(),
		})
	}

	// Return status 200 OK.
	return c.JSON(fiber.Map{
		"error": false,
		"msg":   nil,
		"tokens": fiber.Map{
			"access":  tokens.Access,
			"refresh": tokens.Refresh,
		},
	})
}

// UserSignOut method to de-authorize user and delete refresh token from Redis.
// @Description De-authorize user and delete refresh token from Redis.
// @Summary de-authorize user and delete refresh token from Redis
// @Tags User
// @Accept json
// @Produce json
// @Success 204 {string} status "ok"
// @Security ApiKeyAuth
// @Router /v1/user/sign/out [post]
func UserSignOut(c *fiber.Ctx) error {
	// Get claims from JWT.
	claims, err := utils.ExtractTokenMetadata(c)
	if err != nil {
		// Return status 500 and JWT parse error.
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	// Define user ID.
	userID := strconv.Itoa(claims.UserID)

	// Save refresh token to Redis.
	errDelFromRedis := cache.RedisClient.Del(context.Background(), userID).Err()
	if errDelFromRedis != nil {
		// Return status 500 and Redis deletion error.
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   errDelFromRedis.Error(),
		})
	}

	// Return status 204 no content.
	return c.SendStatus(fiber.StatusNoContent)
}

// GoogleLogin method to generate authenticate url.
// @Description Generate authenticate URL.
// @Summary generate authenticate URL.
// @Tags User
// @Accept json
// @Produce json
// @Success 200 {string} url
// @Router /v1/oauth/google/login [get]
func GoogleLogin(c *fiber.Ctx) error {
	oauthState := utils.GenerateState()
	url := utils.GetAuthCodeURL(oauthState)

	return c.Status(fiber.StatusOK).JSON(fiber.Map{
		"error":            false,
		"msg":              nil,
		"authenticate_url": url,
	})
}

// GoogleCallback method to get user data from Google and create or update user.
// @Description Get data from Google and create/update user.
// @Summary get user data from google.
// @Tags User
// @Accept json
// @Produce json
// @Success 200 {object} models.Token
// @Router /v1/oauth/google/callback [get]
func GoogleCallback(c *fiber.Ctx) error {
	data, err := utils.GetUserDataFromGoogle(c.Query("code"))
	var userData models.GoogleClaims
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}
	err = json.Unmarshal(data, &userData)
	if err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	user := models.User{
		Email:      userData.Email,
		FirstName:  userData.FirstName,
		LastName:   userData.LastName,
		UserRole:   repository.UserRoleName,
		UserStatus: repository.ActiveUserStatus,
		UserImage:  userData.Picture,
	}

	res := database.Database.Clauses(clause.OnConflict{
		Columns:   []clause.Column{{Name: "email"}},
		DoUpdates: clause.AssignmentColumns([]string{"first_name", "last_name", "user_image"}),
	}).Create(&user)
	if res.Error != nil {
		return c.Status(fiber.StatusNotFound).JSON(fiber.Map{
			"error": true,
			"msg":   res.Error.Error(),
		})
	}

	// Get role credentials from founded user.
	credentials, err := utils.GetCredentialsByRole(repository.UserRoleName)
	if err != nil {
		// Return status 400 and error message.
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	// create a JWT for OUR app and give it back to the client for future requests
	tokens, err := utils.GenerateNewTokens(strconv.FormatUint(user.ID, 10), credentials)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   "Couldn't make authentication token",
		})
	}

	// Return status 200 OK.
	return c.JSON(fiber.Map{
		"error": false,
		"msg":   nil,
		"tokens": fiber.Map{
			"access":  tokens.Access,
			"refresh": tokens.Refresh,
		},
	})
}

// ResetPassword method to send email reset password to user.
// @Description Send email reset password.
// @Summary send email reset password.
// @Tags User
// @Accept json
// @Produce json
// @Success 200 {string} url
// @Router /v1/user/reset-password [post]
func ResetPassword(c *fiber.Ctx) error {
	var emailResetPassword models.EmailResetPassword
	if err := c.BodyParser(&emailResetPassword); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	validate := utils.NewValidator()
	if err := validate.Struct(&emailResetPassword); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	var user models.User
	tx := database.Database.First(&user, "email = ?", emailResetPassword.Email)
	if tx.Error != nil {
		return c.Status(fiber.StatusNotFound).JSON(fiber.Map{
			"error": true,
			"msg":   "user not found",
		})
	}

	// Get role credentials from founded user.
	credentials, err := utils.GetCredentialsByRole(user.UserRole)
	if err != nil {
		// Return status 400 and error message.
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	// Generate a new pair of access and refresh tokens.
	tokens, err := utils.GenerateNewTokens(strconv.FormatUint(user.ID, 10), credentials)
	if err != nil {
		// Return status 500 and token generation error.
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	from := os.Getenv("MAIL_ADDRESS")
	password := os.Getenv("MAIL_PASSWORD")
	msg := gomail.NewMessage()
	msg.SetHeader("From", from)
	msg.SetHeader("To", emailResetPassword.Email)
	msg.SetHeader("Subject", "Reset password")
	msg.SetBody("text/html", fmt.Sprintf(repository.MailTemplateResetPassword, os.Getenv("RESET_PASSWORD_CALLBACK_URL")+"?token="+tokens.Access))

	n := gomail.NewDialer("smtp.gmail.com", 587, from, password)

	if err := n.DialAndSend(msg); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}
	return c.JSON(fiber.Map{
		"error": false,
		"msg":   nil,
	})
}

// CreateNewPassword method to create new password for user.
// @Description Create new password for user.
// @Summary create new password.
// @Tags User
// @Accept json
// @Produce json
// @Success 200 {int} id
// @Security ApiKeyAuth
// @Router /v1/user/create-password [post]
func CreateNewPassword(c *fiber.Ctx) error {
	claims, err := utils.ExtractTokenMetadata(c)
	if err != nil {
		return c.Status(fiber.StatusBadGateway).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	var newPassword models.CreatePasswordClaims
	if err := c.BodyParser(&newPassword); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	validate := utils.NewValidator()
	if err := validate.Struct(&newPassword); err != nil {
		// Return, if some fields are not valid.
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   utils.ValidatorErrors(err),
		})
	}

	var user models.User
	tx := database.Database.First(&user, "id = ?", claims.UserID)
	if tx.Error != nil {
		return c.Status(fiber.StatusNotFound).JSON(fiber.Map{
			"error": true,
			"msg":   "user id not found",
		})
	}

	tx = database.Database.Model(&models.User{}).Where("id = ?", claims.UserID).Update("password_hash", utils.GeneratePassword(newPassword.Password))
	if tx.Error != nil {
		return c.Status(fiber.StatusNotFound).JSON(fiber.Map{
			"error": true,
			"msg":   tx.Error.Error(),
		})
	}

	return c.JSON(fiber.Map{
		"error": false,
		"msg":   claims.UserID,
	})
}

func SocialCallback(c *fiber.Ctx) error {
	var userData models.SocialClaims

	if err := c.BodyParser(&userData); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	if err := utils.NewValidator().Struct(userData); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   utils.ValidatorErrors(err),
		})
	}

	if !userData.VerifyExpiresAt(time.Now().Unix(), true) {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   "token expired",
		})
	}

	user := models.User{
		Email:      userData.Email,
		FirstName:  userData.FirstName,
		LastName:   userData.LastName,
		UserRole:   repository.UserRoleName,
		UserStatus: repository.ActiveUserStatus,
		UserImage:  userData.Picture,
	}

	res := database.Database.Clauses(clause.OnConflict{
		Columns:   []clause.Column{{Name: "email"}},
		DoUpdates: clause.AssignmentColumns([]string{"first_name", "last_name", "user_image"}),
	}).Create(&user)
	if res.Error != nil {
		return c.Status(fiber.StatusNotFound).JSON(fiber.Map{
			"error": true,
			"msg":   res.Error.Error(),
		})
	}

	// Get role credentials from founded user.
	credentials, err := utils.GetCredentialsByRole(repository.UserRoleName)
	if err != nil {
		// Return status 400 and error message.
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": true,
			"msg":   err.Error(),
		})
	}

	// create a JWT for OUR app and give it back to the client for future requests
	tokens, err := utils.GenerateNewTokens(strconv.FormatUint(user.ID, 10), credentials)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": true,
			"msg":   "Couldn't make authentication token",
		})
	}

	// Return status 200 OK.
	return c.JSON(fiber.Map{
		"error": false,
		"msg":   nil,
		"tokens": fiber.Map{
			"access":  tokens.Access,
			"refresh": tokens.Refresh,
		},
	})
}